reporter: Refactor template processing out of NoticeTemplateReporter
Move code for processing Freemarker templates from NoticeTemplatReporter
into its own class, to make it more generic and usable for future
reporters that rely on output generated of templates.
