+++ reporter/src/main/resources/templates/notice/summary.ftl
